Create CellTrackingBase, let other classes inherit most of the (otherwise duplicated) methods
